SKALE-1779 Add key name correctness check

+++ SGXWalletServer.cpp
@@ -29,6 +29,19 @@
 #include "SGXWalletServer.h"
 #include "SGXWalletServer.hpp"
 
+#include <algorithm>
+
+bool isStringDec( std::string & str){
+  auto res = std::find_if_not(str.begin(), str.end(), [](char c)->bool{
+    return std::isdigit(c);
+  });
+  return !str.empty() && res == str.end();
+  // bool res =tr
+  // for (int i = 0; i < str.length; i++){
+
+//  }
+}
+
 SGXWalletServer::SGXWalletServer(AbstractServerConnector &connector,
                                  serverVersion_t type)
         : AbstractStubServer(connector, type) {}
@@ -179,7 +192,7 @@ Json::Value generateECDSAKeyImpl() {
     return result;
 }
 
-Json::Value renameESDSAKeyImpl(const std::string& KeyName, const std::string& tempKeyName){
+Json::Value renameECDSAKeyImpl(const std::string& KeyName, const std::string& tempKeyName){
   Json::Value result;
   result["status"] = 0;
   result["errorMessage"] = "";
@@ -189,14 +202,19 @@ Json::Value renameESDSAKeyImpl(const std::string& KeyName, const std::string& te
 
     std::string prefix = tempKeyName.substr(0,8);
     if (prefix != "tmp_NEK:") {
-     throw RPCException(UNKNOWN_ERROR, "");
+     throw RPCException(UNKNOWN_ERROR, "wrong temp key name");
     }
-    prefix = KeyName.substr(0,5);
+    prefix = KeyName.substr(0,12);
     if (prefix != "NEK_NODE_ID:") {
-      throw RPCException(UNKNOWN_ERROR, "");
+      throw RPCException(UNKNOWN_ERROR, "wrong key name");
+    }
+    std::string postfix = KeyName.substr(12, KeyName.length());
+    if (!isStringDec(postfix)){
+      throw RPCException(UNKNOWN_ERROR, "wrong key name");
     }
 
     std::shared_ptr<std::string> key_ptr = readFromDb(tempKeyName);
+    std::cerr << "new key name is " << KeyName <<std::endl;
     writeDataToDB(KeyName, *key_ptr);
     levelDb->deleteTempNEK(tempKeyName);
 
@@ -417,7 +435,7 @@ Json::Value CreateBLSPrivateKeyImpl(const std::string & BLSKeyName, const std::s
     //std::cerr << sshares << std::endl;
     //std::cerr << "length is " << strlen(sshares);
 
-    std::shared_ptr<std::string> encryptedKeyHex_ptr = readFromDb(EthKeyName);//readECDSAKey(EthKeyName);
+    std::shared_ptr<std::string> encryptedKeyHex_ptr = readFromDb(EthKeyName);
 
     bool res = CreateBLSShare(BLSKeyName, sshares, encryptedKeyHex_ptr->c_str());
      if ( res){
@@ -523,9 +541,9 @@ Json::Value SGXWalletServer::generateECDSAKey() {
     return generateECDSAKeyImpl();
 }
 
-Json::Value SGXWalletServer::renameESDSAKey(const std::string& KeyName, const std::string& tempKeyName){
+Json::Value SGXWalletServer::renameECDSAKey(const std::string& KeyName, const std::string& tempKeyName){
   lock_guard<recursive_mutex> lock(m);
-  return renameESDSAKeyImpl(KeyName, tempKeyName);
+  return renameECDSAKeyImpl(KeyName, tempKeyName);
 }
 
 Json::Value SGXWalletServer::getPublicECDSAKey(const std::string &_keyName) {
@@ -668,3 +686,4 @@ void writeDataToDB(const string & Name, const string &value) {
 
   levelDb->writeString(key, value);
 }
+
